numpy => python
